pyproject.toml
[project]
name = "appspider"
version = "0.1.0"
description = "Automated mobile app UI spider — crawls and documents screens, elements, and navigation flows"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.52.0",
    "click>=8.1",
    "Pillow>=11.0",
    "networkx>=3.4",
    "imagehash>=4.3",
]

[project.scripts]
appspider = "appspider.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/appspider"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
